+++ tsc/testdata/baselines/reference/submodule/compiler/jsFileMethodOverloads2.js
@@ -150,5 +150,5 @@ declare class Example<T> {
      * @param {(y: T) => unknown} [fn]
      * @returns {unknown}
      */
-    transform(): T;
+    transform<U>(): T;
 }